Move benchmark to its own package and fix same-type relationship logs/metrics
